tests/wpt/web-platform-tests/shadow-dom/html-elements-and-their-shadow-trees/test-004.html

<!DOCTYPE html>
<!--
Distributed under both the W3C Test Suite License [1] and the W3C
3-clause BSD License [2]. To contribute to a W3C Test Suite, see the
policies and contribution forms [3].

[1] http://www.w3.org/Consortium/Legal/2008/04-testsuite-license
[2] http://www.w3.org/Consortium/Legal/2008/03-bsd-license
[3] http://www.w3.org/2004/10/27-testcases
-->
<html>
<head>
<title>Shadow DOM Test: A_09_00_04</title>
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<link rel="help" href="http://www.w3.org/TR/2013/WD-shadow-dom-20130514/#html-elements-and-their-shadow-trees">
<meta name="assert" content="HTML Elements and Their Shadow Trees: Check that details can contain at least two insertion points with matching criteria 'summary:first-of-type' and 'universal selector'">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script>
//test universal selector
test(unit(function (ctx) {

	var d = newRenderedHTMLDocument(ctx);

	// create element
    var el = d.createElement('details');
    el.setAttribute('open', 'open');
    d.body.appendChild(el);

    el.innerHTML = '' +
    	'<span id="shadow">This is a node that should be distributed</span>' +
    	'<span id="flbk">This is a fallback content</span>';

    var s = el.createShadowRoot();
    s.innerHTML = '<content select="#shadow"></content>';

    assert_true(d.querySelector('#shadow').offsetTop > 0, '\'details\' should allow at least one insertion point');
    assert_equals(d.querySelector('#flbk').offsetTop, 0, 'Fallback content shouldn\'t be rendered');

}), 'A_09_00_04_T01');



//test summary:first-of-type
test(unit(function (ctx) {

	var d = newRenderedHTMLDocument(ctx);

	// create element
    var el = d.createElement('details');
    el.setAttribute('open', 'open');
    d.body.appendChild(el);

    el.innerHTML = '' +
    	'<summary>' +
    	'<span id="shadow">This is a node that should be distributed</span>' +
    	'</summary>' +
    	'<span id="flbk">Unlucky content</span>';

    var s = el.createShadowRoot();
    s.innerHTML = '<content select="summary:first-of-type"></content>';

    assert_true(d.querySelector('#shadow').offsetTop > 0, 'details should allow insertion point' +
    		'with summary:first-of-type matching criteria');
    assert_equals(d.querySelector('#flbk').offsetTop, 0, 'Fallback content shouldn\'t be rendered');

}), 'A_09_00_04_T02');
</script>
</body>
</html>
